requirements.txt
aniso8601==8.1.0
appdirs==1.4.4
appnope==0.1.2
backcall==0.2.0
CacheControl==0.12.6
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.2.1
cycler==0.10.0
decorator==4.4.2
distlib==0.3.1
dnspython==2.0.0
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.9
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
future==0.18.2
gcloud==0.17.0
geocoder==1.38.1
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-cloud-core==1.5.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.35.0
google-crc32c==1.1.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
gunicorn==20.0.4
h5py==3.1.0
httplib2==0.18.1
idna==2.10
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jws==0.1.3
jwt==1.1.0
Keras==2.4.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
msgpack==1.0.2
numpy==1.19.4
oauth2client==3.0.0
pandas==1.2.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pipenv==2020.11.15
prompt-toolkit==3.0.8
proto-plus==1.13.0
protobuf==3.14.0
psycopg2==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.4.3
Pygments==2.7.3
pymongo==3.11.2
pyparsing==2.4.7
Pyrebase==3.0.27
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.15.0
python-jwt==2.0.1
pytz==2020.5
PyYAML==5.3.1
pyzmq==20.0.0
ratelim==0.1.6
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.6
scikit-learn==0.24.0
scipy==1.5.4
six==1.15.0
sklearn==0.0
SQLAlchemy==1.3.22
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.2
virtualenv==20.2.2
virtualenv-clone==0.5.4
Wand==0.6.5
wcwidth==0.2.5
Werkzeug==1.0.1
